feat: Add sorting functionality for student and personnel lists with new utility functions
