+++ package.json
@@ -17,6 +17,7 @@
     "axios": "^1.8.4",
     "cally": "^0.8.0",
     "chart.js": "^4.4.9",
+    "chartjs-plugin-datalabels": "^2.2.0",
     "d3-dsv": "2",
     "date-fns": "^4.1.0",
     "firebase": "^11.6.0",
