+++ src/pages/Admin/ManageAdminRoles.jsx
@@ -5,6 +5,7 @@ import Pagination from "../../components/Pagination";
 import Breadcrumbs from "../../components/Breadcrumbs";
 import { usePersonnelStore } from "../../stores/admin.store"; // ใช้ store ที่สร้างขึ้นมา
 import { useAuthStore } from "../../stores/auth.store";
+import {sortPersonnelOptions, switchSortPersonnel} from "../../utils/sortDataTable"
 
 const ManageAdminRoles = () => {
   // ใช้ Zustand store เพื่อจัดการข้อมูลบุคลากร
@@ -17,14 +18,7 @@ const ManageAdminRoles = () => {
   } = usePersonnelStore();
   const [searchKeyword, setSearchKeyword] = useState("");
   const [selectedOption, setSelectedOption] = useState("SortToMost");
-  const optionsForPersonnel = [
-    { value: "SortToMost", label: "เรียงจากน้อยไปมาก" },
-    { value: "SortToLess", label: "เรียงจากมากไปน้อย" },
-    { value: "AlphaSortToMost", label: "เรียงตามลำดับตัวอักษร ก-ฮ" },
-    { value: "AlphaSortToLess", label: "เรียงตามลำดับตัวอักษร ฮ-ก" },
-    { value: "PrefixMr", label: "คำนำหน้า นาย-นาง" },
-    { value: "PrefixMrs", label: "คำนำหน้า นาง-นาย" },
-  ];
+  
   // สร้าง satate สำหรับ Paginations
   const [currentPage, setCurrentPage] = useState(1);
   const itemsPerPage = 10;
@@ -84,45 +78,7 @@ const ManageAdminRoles = () => {
 
     // เรียงลำดับ
     const sorted = [...searched];
-    switch (selectedOption) {
-      case "SortToMost":
-        sorted.sort((a, b) => a.user_id - b.user_id);
-        break;
-      case "SortToLess":
-        sorted.sort((a, b) => b.user_id - a.user_id);
-        break;
-      case "AlphaSortToMost":
-        sorted.sort((a, b) => {
-          const nameA = `${a.first_name}${a.last_name}`;
-          const nameB = `${b.first_name}${b.last_name}`;
-          return nameA.localeCompare(nameB, "th", { sensitivity: "base" });
-        });
-        break;
-      case "AlphaSortToLess":
-        sorted.sort((a, b) => {
-          const nameA = `${a.first_name}${a.last_name}`;
-          const nameB = `${b.first_name}${b.last_name}`;
-          // { sensitivity: "base" } คือการเปรียบเทียบแบบไม่สนใจตัวพิมพ์ใหญ่-เล็ก
-          // localeCompare() ใช้สำหรับเปรียบเทียบสตริงตามภาษาที่กำหนด
-          return nameB.localeCompare(nameA, "th", { sensitivity: "base" });
-        });
-        break;
-      case "PrefixMr":
-        sorted.sort((a, b) => {
-          const prefixOrder = { นาย: 1, นางสาว: 2, นาง: 3 };
-          return (prefixOrder[a.prefix] || 99) - (prefixOrder[b.prefix] || 99);
-        });
-        break;
-      case "PrefixMrs":
-        sorted.sort((a, b) => {
-          const prefixOrder = { นาย: 1, นางสาว: 2, นาง: 3 };
-          return (prefixOrder[b.prefix] || 99) - (prefixOrder[a.prefix] || 99);
-        });
-
-        break;
-      default:
-        break;
-    }
+    switchSortPersonnel(selectedOption, sorted);
 
     return sorted;
   }, [personnel, searchKeyword, selectedOption]);
@@ -163,7 +119,7 @@ const ManageAdminRoles = () => {
             <FilterDropdown
               selectedOption={selectedOption}
               setSelectedOption={setSelectedOption}
-              options={optionsForPersonnel}
+              options={sortPersonnelOptions}
               className="select"
             />
           </div>
